Fix #14 - format negative zero as zero.
Negative zero is a construct of IEEE 754 floating point, not a mathematical
concept that (most) humans should care about. And since you may see tick values
that are very small negative numbers due to limited precision, such as -8e-17,
it’s helpful to format these as zero (“0.00”), not negative zero (“-0.00”).