export locales without renaming
This change updates the locale exports to not automatically append the         
`.json` suffix. The export pattern is a simple matching pattern as            
described in https://nodejs.org/api/packages.html#subpath-patterns and        
while friendly, it prohibits TypeScript from consuming the locales with        
its `resolveJsonModule` option which expect the import to have the            
`.json` suffix. Because the export is helpfully appending the `.json`         
suffix this results in a lookup of `./locale/<locale>.json.json` which
will then fail.